+++ cvat-ui/postcss.config.js
@@ -1,0 +1,9 @@
+/* eslint-disable */
+module.exports = {
+    parser: false,
+    plugins: {
+        'postcss-preset-env': {
+            browsers: '> 2%', // https://github.com/browserslist/browserslist
+        },
+    },
+};