+++ cvat-ui/src/components/annotation-page/standard-workspace/canvas-wrapper-component.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import {
+    Layout,
+} from 'antd';
+
+export default function CanvasWrapperComponent(): JSX.Element {
+    return (
+        <Layout.Content
+            className='cvat-annotation-page-canvas-container'
+        >
+            main content
+        </Layout.Content>
+    );
+}