+++ cvat/apps/engine/static/engine/js/shapes.js
@@ -3014,7 +3014,8 @@ class PolyShapeView extends ShapeView {
 
                             // Run edit mode
                             PolyShapeView.editor.edit(this._controller.type.split('_')[1],
-                                this._uis.shape.attr('points'), this._color, index, e,
+                                this._uis.shape.attr('points'), this._color, index,
+                                this._uis.shape.attr('points').split(/\s/)[index], e,
                                 (points) => {
                                     this._uis.shape.removeClass('hidden');
                                     if (this._uis.points) {
@@ -3024,7 +3025,8 @@ class PolyShapeView extends ShapeView {
                                         this._uis.shape.attr('points', points);
                                         this._controller.updatePosition(window.cvat.player.frames.current, this._buildPosition());
                                     }
-                                }
+                                },
+                                this._controller.id
                             );
                         }
                     }